+++ docs/releases.md
@@ -5,8 +5,9 @@ The changes have been tracked starting version 0.1.0
 
 ## version 0.2.7
 
+* adding `compareReportSuites` method
 * adding `reportType` as attribute from projects. Either `desktop` or `mobile`
-* 
+* adding `version` and `curation` attributes by default on dict version of `project`.
 
 ## version 0.2.6